requirements.txt
reflex==0.4.9

Flask~=3.0.3
PyYAML~=6.0.1
yfinance~=0.2.38